Scripts to run the archive in the GRPC container

+++ tests/runDetectSugars/scripts/ontology.sh
@@ -57,65 +57,51 @@ run_PDB() {
 	FILENAME="$(echo "${ZIPFILE}" | cut -c4-7)"
 
 
-#if grep -q "$FILENAME" 12-13-20/12-13-20LoopKilled.txt; then
+	# If the file is in the Updated PDB List (AKA the new files rsync copied since last time), then run script.
+	# -q stops grep from writing output
+	if grep -qi "$FILENAME" UpdatedPDBs.txt; then
 
-#If the file is in the Updated PDB List (AKA the new files rsync copied since last time), then run script.
-#if grep -qi "$FILENAME" UpdatedPDBs.txt; then
+		echo "FILE: ${FILEPATH}"
+		echo "ZIPFILE: ${ZIPFILE}"
+		echo "FILENAME: ${FILENAME}"
+		cp "${FILEPATH}" "${DEST}"
 
-	echo "FILE: ${FILEPATH}"
-	echo "ZIPFILE: ${ZIPFILE}"
-	echo "FILENAME: ${FILENAME}"
-	cp "${FILEPATH}" "${DEST}"
+		## Set the output variables.
+		ZIPDEST="$(echo "${DEST}/${ZIPFILE}")"
+		ENTDEST="$(echo "${DEST}/pdb${FILENAME}.ent")"
+		FILENAME="$(echo "${FILENAME}" | tr [:lower:] [:upper:])"
+		FILEDEST="$(echo "${PDBDEST}/${FILENAME}.pdb")"
 
-	## Set the output variables.
-	ZIPDEST="$(echo "${DEST}/${ZIPFILE}")"
-	ENTDEST="$(echo "${DEST}/pdb${FILENAME}.ent")"
-	FILENAME="$(echo "${FILENAME}" | tr [:lower:] [:upper:])"
-	FILEDEST="$(echo "${PDBDEST}/${FILENAME}.pdb")"
+		## Unzip he gzipped file and move it to it's destination.
+		gunzip "${ZIPDEST}"
+		mv "${ENTDEST}" "${FILEDEST}"
+		REPDEST="$(echo "${RDEST}/${FILENAME}.txt")"
 
-	## Unzip he gzipped file and move it to it's destination.
-	gunzip "${ZIPDEST}"
-	mv "${ENTDEST}" "${FILEDEST}"
-	REPDEST="$(echo "${RDEST}/${FILENAME}.txt")"
+		## Print some variables for debugging purposes.
+		echo "FILEDEST: ${FILEDEST}"
+		echo "REPDEST: ${REPDEST}"
 
-	## Print some variables for debugging purposes.
-	echo "FILEDEST: ${FILEDEST}"
-	echo "REPDEST: ${REPDEST}"
+		# Run the pdb file through detect_sugars
 
-	#python3 ${GEMSHOME}/testbin/cycle.py -amino_libs ${AMINO_LIBS} -pdb ${FILEDEST} &> ${REPDEST} & pid=$!
-        
-	
-#IF GREP $FILENAME is not in 12-7-20.log then run script	
-#-q stops grep from writing output
-#	if ! grep -q "$FILENAME" 12-7-20.log; then
-
-#If the file is in the Updated PDB List (AKA the new files rsync copied since last time), then run script.
-#	if grep -qi "$FILENAME" UpdatedPDBs.txt; then
-		${GEMSHOME}/gmml/tests/archiveRun_detect_sugars ${FILEDEST} &> ${REPDEST} 2>&1 && rm -f ${FILEDEST} >> /dev/null 2>&1 & pid=$!
-
-	#( sleep 60 && kill -HUP $pid ) 2>/dev/null & watcher=$!
-	#if wait $pid 2>/dev/null; then
-        #   pkill -HUP -P $watcher
-        #   wait $watcher
-        #else
-        #   printf "Failed\n"
-        #   echo ${FILEDEST} >> killed.txt
-        #fi
-
-	## Some of the PDBs cause our code to get stuck in forever loops.
-	## This will kill any straying processes, so the overall purpose of the
-	## script can continue. This will log an Error, so we can go back and figure
-	## out why.
+		#python3 ${GEMSHOME}/testbin/cycle.py -amino_libs ${AMINO_LIBS} -pdb ${FILEDEST} &> ${REPDEST} & pid=$!
+			
+		${GEMSHOME}/gmml/tests/archiveRun_detect_sugars ${FILEDEST} &> ${REPDEST} 2>&1 && \
+			rm -f ${FILEDEST} >> /dev/null 2>&1 & pid=$!
+		
+		# Some of the PDBs cause our code to get stuck in infinite loops.
+		# This will kill any straying processes, so the overall purpose of the
+		# script can continue. This will log an Error, so we can go back and figure
+		# out why.
 		time=1
 		while kill -0 ${pid} > /dev/null 2>&1; do
 			if [ ${time} -gt 4800000 ]; then
 				kill -9 ${pid}
 				sleep 5
 				echo "${FILENAME}" >> killed.txt
- 			fi
+			fi
 			((time++))
 		done
-#fi
+	fi
 } ## End function run_PDB
 
 ## MAIN
